requirements.txt
# requirements.txt
click>=8.0.0
PyYAML>=6.0
rich>=10.0.0
setuptools
